+++ src/lib/components/atoms/index.ts
@@ -2,4 +2,5 @@ export * from './Alert';
 export * from './Button';
 export * from './Progress';
 export * from './Section';
+export * from './Toast';
 export * from './Tooltip';
