+++ src/lib/components/index.ts
@@ -21,7 +21,7 @@ export * from './Spinner';
 export * from './Tab';
 export * from './Table';
 export * from './Timeline';
-export * from './Toast';
+
 
 export type DeepPartial<T> = T extends object
   ? {
