+++ src/lib/theme/default.ts
@@ -3,6 +3,7 @@ import { ALERT_THEME } from '../components/atoms/Alert/Alert.theme';
 import { BUTTON_THEME } from '../components/atoms/Button/Button.theme';
 import { PROGRESS_THEME } from '../components/atoms/Progress/Progress.theme';
 import { SECTION_THEME } from '../components/atoms/Section/Section.theme';
+import { TOAST_THEME } from '../components/atoms/Toast/Toast.theme';
 import { TOOLTIP_THEME } from '../components/atoms/Tooltip/Tooltip.theme';
 
 const theme: HelloInternetTheme = {
@@ -717,15 +718,7 @@ const theme: HelloInternetTheme = {
     },
     tabpanel: 'p-4'
   },
-  toast: {
-    base: 'flex w-full max-w-xs items-center rounded-lg bg-white p-4 text-gray-500 shadow dark:bg-gray-800 dark:text-gray-400',
-    closed: 'opacity-0 ease-out',
-    removed: 'hidden',
-    toggle: {
-      base: '-mx-1.5 -my-1.5 ml-auto inline-flex h-8 w-8 rounded-lg bg-white p-1.5 text-gray-400 hover:bg-gray-100 hover:text-gray-900 focus:ring-2 focus:ring-gray-300 dark:bg-gray-800 dark:text-gray-500 dark:hover:bg-gray-700 dark:hover:text-white',
-      icon: 'h-5 w-5 shrink-0'
-    }
-  },
+  toast: TOAST_THEME,
   tooltip: TOOLTIP_THEME
 };
